hides info about the team
after teams submit their texts they receive 10 texts from other teams which they'll anonymously read and grade (a sort of nationwide peer-review!). the thing is that many teams try to find each other by their team names and city of origin (you'll find past examples of this on twitter, although it's more common in private groups in whatsapp). anyway, this code is intend to be active only when they're grading each other, an attempt to reduce this effect